+++ libindy/src/domain/crypto/key.rs
@@ -29,17 +29,17 @@ impl Key {
 
 //Can't figure out why this macro can't be found in `cargo test` because
 //if works for src/utils/crypto/randombytes/sodium
-//memzeroize!(Key, signkey);
+memzeroize!(Key, signkey);
 
-impl Zeroize for Key {
-    fn zeroize(&mut self) { self.signkey.zeroize(); }
-}
-
-impl Drop for Key {
-    fn drop(&mut self) {
-        self.signkey.zeroize();
-    }
-}
+//impl Zeroize for Key {
+//    fn zeroize(&mut self) { self.signkey.zeroize(); }
+//}
+//
+//impl Drop for Key {
+//    fn drop(&mut self) {
+//        self.signkey.zeroize();
+//    }
+//}
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct KeyInfo {
